refactor: split session control modules
