+++ services/api-rs/crates/centaur-session-core/src/metadata.rs
@@ -1,0 +1,5 @@
+use serde_json::Value;
+
+pub fn empty_object() -> Value {
+    Value::Object(serde_json::Map::new())
+}
