+++ services/api-rs/crates/centaur-session-core/src/thread_key.rs
@@ -1,0 +1,134 @@
+use std::{fmt, str::FromStr};
+
+use serde::{Deserialize, Deserializer, Serialize, Serializer, de};
+use thiserror::Error;
+
+pub const MAX_THREAD_KEY_BYTES: usize = 512;
+
+#[derive(Clone, Debug, Eq, PartialEq, Hash)]
+pub struct ThreadKey(String);
+
+impl ThreadKey {
+    pub fn parse(value: impl Into<String>) -> Result<Self, ThreadKeyError> {
+        let value = value.into();
+        validate_thread_key(&value)?;
+        Ok(Self(value))
+    }
+
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+
+    pub fn into_string(self) -> String {
+        self.0
+    }
+}
+
+impl fmt::Display for ThreadKey {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str(self.as_str())
+    }
+}
+
+impl FromStr for ThreadKey {
+    type Err = ThreadKeyError;
+
+    fn from_str(value: &str) -> Result<Self, Self::Err> {
+        Self::parse(value)
+    }
+}
+
+impl TryFrom<String> for ThreadKey {
+    type Error = ThreadKeyError;
+
+    fn try_from(value: String) -> Result<Self, Self::Error> {
+        Self::parse(value)
+    }
+}
+
+impl AsRef<str> for ThreadKey {
+    fn as_ref(&self) -> &str {
+        self.as_str()
+    }
+}
+
+impl Serialize for ThreadKey {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_str(self.as_str())
+    }
+}
+
+impl<'de> Deserialize<'de> for ThreadKey {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let value = String::deserialize(deserializer)?;
+        Self::parse(value).map_err(de::Error::custom)
+    }
+}
+
+#[derive(Clone, Debug, Eq, PartialEq, Error)]
+pub enum ThreadKeyError {
+    #[error("thread_key is required")]
+    Empty,
+    #[error("thread_key must be at most {MAX_THREAD_KEY_BYTES} bytes")]
+    TooLong,
+    #[error("thread_key must be namespaced as '<source>:<id>'")]
+    MissingNamespace,
+    #[error("thread_key must not contain ASCII control characters")]
+    ControlCharacter,
+    #[error("thread_key must not be raw JSON")]
+    RawJson,
+}
+
+fn validate_thread_key(value: &str) -> Result<(), ThreadKeyError> {
+    if value.is_empty() {
+        return Err(ThreadKeyError::Empty);
+    }
+    if value.len() > MAX_THREAD_KEY_BYTES {
+        return Err(ThreadKeyError::TooLong);
+    }
+    if value.starts_with('{') || value.starts_with('[') {
+        return Err(ThreadKeyError::RawJson);
+    }
+    if value.chars().any(|ch| ch.is_ascii_control()) {
+        return Err(ThreadKeyError::ControlCharacter);
+    }
+    let Some((namespace, rest)) = value.split_once(':') else {
+        return Err(ThreadKeyError::MissingNamespace);
+    };
+    if namespace.is_empty() || rest.is_empty() {
+        return Err(ThreadKeyError::MissingNamespace);
+    }
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::ThreadKey;
+
+    #[test]
+    fn thread_key_accepts_namespaced_values() {
+        let key = ThreadKey::parse("chat:C123:1780000000.000000").unwrap();
+        assert_eq!(key.as_str(), "chat:C123:1780000000.000000");
+    }
+
+    #[test]
+    fn thread_key_rejects_missing_namespace() {
+        let err = ThreadKey::parse("not-namespaced").unwrap_err();
+        assert_eq!(
+            err.to_string(),
+            "thread_key must be namespaced as '<source>:<id>'"
+        );
+    }
+
+    #[test]
+    fn thread_key_rejects_unbounded_payload_shape() {
+        let err = ThreadKey::parse("{\"thread\":\"x\"}").unwrap_err();
+        assert_eq!(err.to_string(), "thread_key must not be raw JSON");
+    }
+}
